+++ maya/AbcImport/AlembicNode.cpp
@@ -299,7 +299,8 @@ MStatus AlembicNode::compute(const MPlug & plug, MDataBlock & dataBlock)
 
         mOutRead[0] = true;
 
-        unsigned int propSize = mData.mPropList.size();
+        unsigned int propSize =
+            static_cast<unsigned int>(mData.mPropList.size());
 
         if (propSize > 0)
         {
@@ -357,7 +358,8 @@ MStatus AlembicNode::compute(const MPlug & plug, MDataBlock & dataBlock)
 
         mOutRead[1] = true;
 
-        unsigned int xformSize = mData.mXformList.size();
+        unsigned int xformSize =
+            static_cast<unsigned int>(mData.mXformList.size());
 
         if (xformSize > 0)
         {
@@ -414,7 +416,8 @@ MStatus AlembicNode::compute(const MPlug & plug, MDataBlock & dataBlock)
 
         mOutRead[8] = true;
 
-        unsigned int locSize = mData.mLocList.size();
+        unsigned int locSize =
+            static_cast<unsigned int>(mData.mLocList.size());
 
         if (locSize > 0)
         {
@@ -461,7 +464,8 @@ MStatus AlembicNode::compute(const MPlug & plug, MDataBlock & dataBlock)
 
         mOutRead[2] = true;
 
-        unsigned int subDSize = mData.mSubDList.size();
+        unsigned int subDSize =
+            static_cast<unsigned int>(mData.mSubDList.size());
 
         if (subDSize > 0)
         {
@@ -535,7 +539,8 @@ MStatus AlembicNode::compute(const MPlug & plug, MDataBlock & dataBlock)
 
         mOutRead[3] = true;
 
-        unsigned int polySize = mData.mPolyMeshList.size();
+        unsigned int polySize =
+            static_cast<unsigned int>(mData.mPolyMeshList.size());
 
         if (polySize > 0)
         {
@@ -609,7 +614,8 @@ MStatus AlembicNode::compute(const MPlug & plug, MDataBlock & dataBlock)
 
         mOutRead[4] = true;
 
-        unsigned int cameraSize = mData.mCameraList.size();
+        unsigned int cameraSize =
+            static_cast<unsigned int>(mData.mCameraList.size());
 
         if (cameraSize > 0)
         {
@@ -681,7 +687,8 @@ MStatus AlembicNode::compute(const MPlug & plug, MDataBlock & dataBlock)
 
         mOutRead[5] = true;
 
-        unsigned int nSurfaceSize = mData.mNurbsList.size();
+        unsigned int nSurfaceSize =
+            static_cast<unsigned int>(mData.mNurbsList.size());
 
         if (nSurfaceSize > 0)
         {
@@ -719,7 +726,8 @@ MStatus AlembicNode::compute(const MPlug & plug, MDataBlock & dataBlock)
 
         mOutRead[6] = true;
 
-        unsigned int nCurveGrpSize = mData.mCurvesList.size();
+        unsigned int nCurveGrpSize =
+            static_cast<unsigned int>(mData.mCurvesList.size());
 
         if (nCurveGrpSize > 0)
         {
